requirements.txt
urwid==1.3.0
PyYAML==3.11
